Add posterior probabilities as output to `expectation_maximization_registration.register()`
Add `self.P`  as return variable in `expectation_maximization_registration.register()`. This was
done because a main outcome of the CPD algorithm is to identify correspondences between points
in the two meshes. The correspondences for each point in one mesh can be found in the other mesh
by running an argmax (e.g., `np.argmax(self.P, axis=axis`,  where axis is 0 or 1, depending on the
mesh you want correspondences for).

The current implementation is the simplest. However, this will likely break existing code as there
will be an additional output parameter that wasn't present before. Other implementations may be
desired for this reason.